simple-agent/pyproject.toml
[project]
name = "simple-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=0.3.25",
    "langchain-ollama>=0.3.3",
    "openai>=1.79.0",
    "python-dotenv>=1.1.0",
]
